refactor: 모델 이전
