+++ src/interfaces/models/stackTrace.ts
